+++ pkg/api/query.go
@@ -6,31 +6,33 @@ package api
 
 import (
 	"context"
-	"github.com/prometheus/client_golang/prometheus"
-	pgMetrics "github.com/timescale/promscale/pkg/pgmodel/metrics"
 	"net/http"
 	"time"
 
 	"github.com/NYTimes/gziphandler"
+
 	"github.com/timescale/promscale/pkg/log"
 	"github.com/timescale/promscale/pkg/promql"
 )
 
-func Query(conf *Config, queryEngine *promql.Engine, queryable promql.Queryable, metrics *Metrics) http.Handler {
-	hf := corsWrapper(conf, queryHandler(queryEngine, queryable, metrics))
+func Query(conf *Config, queryEngine *promql.Engine, queryable promql.Queryable, updateMetrics func(handler, code string, duration float64)) http.Handler {
+	hf := corsWrapper(conf, queryHandler(queryEngine, queryable, updateMetrics))
 	return gziphandler.GzipHandler(hf)
 }
 
-func queryHandler(queryEngine *promql.Engine, queryable promql.Queryable, metrics *Metrics) http.HandlerFunc {
+func queryHandler(queryEngine *promql.Engine, queryable promql.Queryable, updateMetrics func(handler, code string, duration float64)) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		statusCode := "400"
+		begin := time.Now()
+		defer func() {
+			updateMetrics("/api/v1/query", statusCode, time.Since(begin).Seconds())
+		}()
 		var ts time.Time
 		var err error
 		ts, err = parseTimeParam(r, "time", time.Now())
 		if err != nil {
 			log.Error("msg", "Query error", "err", err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 
@@ -41,8 +43,6 @@ func queryHandler(queryEngine *promql.Engine, queryable promql.Queryable, metric
 			if err != nil {
 				log.Error("msg", "Query error", "err", err.Error())
 				respondError(w, http.StatusBadRequest, err, "bad_data")
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "400"}).Inc()
-				metrics.InvalidQueryReqs.Add(1)
 				return
 			}
 
@@ -50,40 +50,35 @@ func queryHandler(queryEngine *promql.Engine, queryable promql.Queryable, metric
 			defer cancel()
 		}
 
-		begin := time.Now()
 		qry, err := queryEngine.NewInstantQuery(queryable, r.FormValue("query"), ts)
 		if err != nil {
 			log.Error("msg", "Query error", "err", err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "400"}).Inc()
 			return
 		}
 
 		res := qry.Exec(ctx)
-		pgMetrics.QueryDuration.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query"}).Observe(time.Since(begin).Seconds())
-
 		if res.Err != nil {
 			log.Error("msg", res.Err, "endpoint", "query")
 			switch res.Err.(type) {
 			case promql.ErrQueryCanceled:
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "503"}).Inc()
+				statusCode = "503"
 				respondError(w, http.StatusServiceUnavailable, res.Err, "canceled")
 				return
 			case promql.ErrQueryTimeout:
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "503"}).Inc()
+				statusCode = "503"
 				respondError(w, http.StatusServiceUnavailable, res.Err, "timeout")
 				return
 			case promql.ErrStorage:
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "500"}).Inc()
+				statusCode = "500"
 				respondError(w, http.StatusInternalServerError, res.Err, "internal")
 				return
 			}
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "422"}).Inc()
+			statusCode = "422"
 			respondError(w, http.StatusUnprocessableEntity, res.Err, "execution")
 			return
 		}
-		pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query", "code": "2xx"}).Inc()
-
+		statusCode = "2xx"
 		respondQuery(w, res, res.Warnings)
 	}
 }