+++ pkg/telemetry/promql.go
@@ -60,39 +60,39 @@ func getValue(result *promql.Result) (float64, error) {
 var promqlStats = []promqlTelemetry{
 	{
 		name:  "promql_query_execution_time_p50",
-		query: "histogram_quantile(0.5, sum by(le) (rate(promscale_metrics_query_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.5, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler=~"/api/v1/query.*"}[1h])))`,
 	}, {
 		name:  "promql_query_execution_time_p90",
-		query: "histogram_quantile(0.9, sum by(le) (rate(promscale_metrics_query_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.9, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler=~"/api/v1/query.*"}[1h])))`,
 	}, {
 		name:  "promql_query_execution_time_p95",
-		query: "histogram_quantile(0.95, sum by(le) (rate(promscale_metrics_query_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.95, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler=~"/api/v1/query.*"}[1h])))`,
 	}, {
 		name:  "promql_query_execution_time_p99",
-		query: "histogram_quantile(0.99, sum by(le) (rate(promscale_metrics_query_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.99, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler=~"/api/v1/query.*"}[1h])))`,
 	}, {
 		name:  "promql_query_remote_read_batch_execution_time_p50",
-		query: "histogram_quantile(0.5, sum by(le) (rate(promscale_metrics_query_remote_read_batch_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.5, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler="/read"}[1h])))`,
 	}, {
 		name:  "promql_query_remote_read_batch_execution_time_p90",
-		query: "histogram_quantile(0.9, sum by(le) (rate(promscale_metrics_query_remote_read_batch_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.9, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler="/read"}[1h])))`,
 	}, {
 		name:  "promql_query_remote_read_batch_execution_time_p95",
-		query: "histogram_quantile(0.95, sum by(le) (rate(promscale_metrics_query_remote_read_batch_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.95, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler="/read"}[1h])))`,
 	}, {
 		name:  "promql_query_remote_read_batch_execution_time_p99",
-		query: "histogram_quantile(0.99, sum by(le) (rate(promscale_metrics_query_remote_read_batch_duration_seconds_bucket[1h])))",
+		query: `histogram_quantile(0.99, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="metric", handler="/read"}[1h])))`,
 	}, {
 		name:  "trace_query_execution_time_p50",
-		query: `histogram_quantile(0.5, sum by(le) (rate(promscale_query_duration_seconds_bucket{handler="Find_Traces"}[1h])))`,
+		query: `histogram_quantile(0.5, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="trace", handler="Find_Traces"}[1h])))`,
 	}, {
 		name:  "trace_query_execution_time_p90",
-		query: `histogram_quantile(0.9, sum by(le) (rate(promscale_query_duration_seconds_bucket{handler="Find_Traces"}[1h])))`,
+		query: `histogram_quantile(0.9, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="trace", handler="Find_Traces"}[1h])))`,
 	}, {
 		name:  "trace_query_execution_time_p95",
-		query: `histogram_quantile(0.95, sum by(le) (rate(promscale_query_duration_seconds_bucket{handler="Find_Traces"}[1h])))`,
+		query: `histogram_quantile(0.95, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="trace", handler="Find_Traces"}[1h])))`,
 	}, {
 		name:  "trace_query_execution_time_p99",
-		query: `histogram_quantile(0.99, sum by(le) (rate(promscale_query_duration_seconds_bucket{handler="Find_Traces"}[1h])))`,
+		query: `histogram_quantile(0.99, sum by(le) (rate(promscale_query_duration_seconds_bucket{type="trace", handler="Find_Traces"}[1h])))`,
 	},
 }